fix die to renaming

+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/model/objects/OutputObject.java
@@ -1,19 +1,19 @@
 package com.github.dedis.popstellar.model.objects;
 
 public class OutputObject {
-  private final int value;
+  private final long value;
   private final ScriptOutputObject script;
 
   /**
    * @param value the value of the output transaction, expressed in miniLAOs
    * @param script The script describing the TxOut unlock mechanism
    */
-  public OutputObject(int value, ScriptOutputObject script) {
+  public OutputObject(long value, ScriptOutputObject script) {
     this.value = value;
     this.script = script;
   }
 
-  public int getValue() {
+  public long getValue() {
     return value;
   }
