+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/model/objects/TransactionObject.java
@@ -176,14 +176,14 @@ public class TransactionObject {
    * @param receiver Public Key of a potential receiver
    * @return int amount of Lao Coin
    */
-  public int getMiniLaoPerReceiver(PublicKey receiver) {
+  public long getMiniLaoPerReceiver(PublicKey receiver) {
     // Check in the future if useful
     if (!isReceiver(receiver)) {
       throw new IllegalArgumentException(
           "The public Key is not contained in the receiver public key");
     }
     // Set the return value to nothing
-    int miniLao = 0;
+    long miniLao = 0;
     // Compute the hash of the public key
     String hash_key = receiver.computeHash();
     // iterate through the output and sum if it's for the argument public key
